fix: address schema validation review feedback
